Breaking: Move PublicationFieldTypes into Concerns namespace
After conferring with ChatGPT I think this makes the most semantic sense, as it defines shared logic and data for multiple domains, while not being a "true" concern, the Concerns namespace feels like the least bad of options available.
