+++ packages/publications/src/Actions/PublicationFieldValidator.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace Hyde\Publications\Actions;
 
-use Hyde\Publications\PublicationService;
-use Hyde\Publications\PublicationFieldTypes;
-use Hyde\Publications\Models\PublicationType;
-use Illuminate\Contracts\Validation\Validator;
+use Hyde\Publications\Concerns\PublicationFieldTypes;
 use Hyde\Publications\Models\PublicationFieldDefinition;
+use Hyde\Publications\Models\PublicationType;
+use Hyde\Publications\PublicationService;
+use Illuminate\Contracts\Validation\Validator;
 
+use function array_merge;
 use function collect;
 use function validator;
-use function array_merge;
 
 /**
  * @see \Hyde\Publications\Testing\Feature\PublicationFieldValidatorTest
