+++ packages/publications/src/Commands/MakePublicationTypeCommand.php
@@ -5,24 +5,24 @@ declare(strict_types=1);
 namespace Hyde\Publications\Commands;
 
 use Hyde\Hyde;
-use Hyde\Publications\Actions\CreatesNewPublicationType;
-use Hyde\Publications\Models\PublicationFieldDefinition;
-use Hyde\Publications\Models\PublicationTags;
-use Hyde\Publications\PublicationFieldTypes;
-use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use InvalidArgumentException;
+use Illuminate\Support\Collection;
 use LaravelZero\Framework\Commands\Command;
+use Hyde\Publications\Models\PublicationTags;
+use Hyde\Publications\Concerns\PublicationFieldTypes;
+use Hyde\Publications\Actions\CreatesNewPublicationType;
+use Hyde\Publications\Models\PublicationFieldDefinition;
 
-use function array_keys;
+use function trim;
 use function count;
-use function in_array;
 use function is_dir;
 use function is_file;
 use function scandir;
 use function sprintf;
+use function in_array;
+use function array_keys;
 use function strtolower;
-use function trim;
 
 /**
  * Hyde Command to create a new publication type.
