+++ packages/publications/tests/Feature/CreatesNewPublicationPageTest.php
@@ -4,20 +4,20 @@ declare(strict_types=1);
 
 namespace Hyde\Publications\Testing\Feature;
 
-use function file_get_contents;
-
-use Hyde\Facades\Filesystem;
 use Hyde\Hyde;
-use Hyde\Publications\Actions\CreatesNewPublicationPage;
-use Hyde\Publications\Models\PublicationFieldValue;
-use Hyde\Publications\Models\PublicationType;
-use Hyde\Publications\PublicationFieldTypes;
-use Hyde\Testing\TestCase;
-use Illuminate\Support\Carbon;
-use Illuminate\Support\Collection;
-use Illuminate\Support\Str;
 use RuntimeException;
+use Hyde\Testing\TestCase;
+use Illuminate\Support\Str;
+use Hyde\Facades\Filesystem;
+use Illuminate\Support\Carbon;
 use Symfony\Component\Yaml\Yaml;
+use Illuminate\Support\Collection;
+use Hyde\Publications\Models\PublicationType;
+use Hyde\Publications\Models\PublicationFieldValue;
+use Hyde\Publications\Concerns\PublicationFieldTypes;
+use Hyde\Publications\Actions\CreatesNewPublicationPage;
+
+use function file_get_contents;
 
 /**
  * @covers \Hyde\Publications\Actions\CreatesNewPublicationPage
