+++ packages/publications/tests/Feature/StaticSiteBuilderPublicationModuleTest.php
@@ -4,18 +4,17 @@ declare(strict_types=1);
 
 namespace Hyde\Publications\Testing\Feature;
 
-use function collect;
-
-use Hyde\Facades\Filesystem;
 use Hyde\Hyde;
-use Hyde\Publications\Actions\CreatesNewPublicationType;
-use Hyde\Publications\Actions\SeedsPublicationFiles;
-use Hyde\Publications\Models\PublicationType;
-use Hyde\Publications\PublicationFieldTypes;
 use Hyde\Testing\TestCase;
+use Hyde\Facades\Filesystem;
 use Illuminate\Support\Collection;
+use Hyde\Publications\Models\PublicationType;
+use Hyde\Publications\Actions\SeedsPublicationFiles;
+use Hyde\Publications\Concerns\PublicationFieldTypes;
+use Hyde\Publications\Actions\CreatesNewPublicationType;
 
 use function range;
+use function collect;
 
 /**
  * Tests that publication pages are compiled properly when building the static site.
